llmmllab-runner/pyproject.toml
[project]
name = "llmmllab-runner"
version = "0.1.0"
description = "Standalone llama.cpp server manager and request proxy"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "nvsmi>=0.4.0",
]

[tool.uv]
dev-dependencies = []
